Merge develop into new-api

+++ catroid/src/androidTest/java/org/catrobat/catroid/uiespresso/ui/fragment/ProjectOptionsTest.java
@@ -25,9 +25,10 @@ package org.catrobat.catroid.uiespresso.ui.fragment;
 
 import android.app.Activity;
 import android.app.Instrumentation;
+import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.Environment;
+import android.os.Build;
 import android.widget.EditText;
 
 import org.catrobat.catroid.ProjectManager;
@@ -40,11 +41,11 @@ import org.catrobat.catroid.io.StorageOperations;
 import org.catrobat.catroid.io.asynctask.ProjectSaver;
 import org.catrobat.catroid.ui.ProjectActivity;
 import org.catrobat.catroid.uiespresso.util.UiTestUtils;
+import org.catrobat.catroid.uiespresso.util.actions.CustomActions;
 import org.catrobat.catroid.uiespresso.util.rules.FragmentActivityTestRule;
 import org.hamcrest.Matcher;
 import org.hamcrest.core.AllOf;
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -66,6 +67,8 @@ import static org.catrobat.catroid.R.id.tab_layout;
 import static org.catrobat.catroid.common.Constants.CATROBAT_EXTENSION;
 import static org.catrobat.catroid.common.Constants.EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY;
 import static org.catrobat.catroid.common.FlavoredConstants.DEFAULT_ROOT_DIRECTORY;
+import static org.catrobat.catroid.uiespresso.ui.fragment.rvutils.RecyclerViewInteractionWrapper.onRecyclerView;
+import static org.catrobat.catroid.uiespresso.util.UiTestUtils.onToast;
 import static org.catrobat.catroid.uiespresso.util.actions.TabActionsKt.selectTabAtPosition;
 import static org.catrobat.catroid.uiespresso.util.matchers.BundleMatchers.bundleHasExtraIntent;
 import static org.catrobat.catroid.uiespresso.util.matchers.BundleMatchers.bundleHasMatchingString;
@@ -74,6 +77,9 @@ import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.instanceOf;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
 import static androidx.test.espresso.Espresso.closeSoftKeyboard;
 import static androidx.test.espresso.Espresso.onView;
@@ -93,6 +99,7 @@ import static androidx.test.espresso.intent.matcher.IntentMatchers.hasType;
 import static androidx.test.espresso.matcher.RootMatchers.isDialog;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.isNotChecked;
+import static androidx.test.espresso.matcher.ViewMatchers.isRoot;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
@@ -104,7 +111,9 @@ public class ProjectOptionsTest {
 	private static final String EXISTING_PROJECT_NAME = "existingProjectName";
 	private static final String DESCRIPTION = "myDescription";
 	private static final String NOTES_AND_CREDITS = "myNotesAndCredits";
+	private static final Integer DURATION_WAIT_FOR_ZIP_FILE_IN_MILLISECONDS = 3000;
 	private static Project project = null;
+	private static Context context = null;
 
 	@Rule
 	public FragmentActivityTestRule<ProjectActivity> baseActivityTestRule =
@@ -113,10 +122,11 @@ public class ProjectOptionsTest {
 
 	@Before
 	public void setUp() throws Exception {
+		context = ApplicationProvider.getApplicationContext();
 		project = UiTestUtils.createEmptyProject(EXISTING_PROJECT_NAME);
-		new ProjectSaver(project, ApplicationProvider.getApplicationContext()).saveProjectAsync();
+		new ProjectSaver(project, context).saveProjectAsync();
 		project = UiTestUtils.createEmptyProject(PROJECT_NAME);
-		new ProjectSaver(project, ApplicationProvider.getApplicationContext()).saveProjectAsync();
+		new ProjectSaver(project, context).saveProjectAsync();
 		baseActivityTestRule.launchActivity(null);
 
 		openContextualActionModeOverflowMenu();
@@ -145,16 +155,16 @@ public class ProjectOptionsTest {
 		closeSoftKeyboard();
 
 		File projectFile = new File(DEFAULT_ROOT_DIRECTORY, PROJECT_NAME);
-		Assert.assertTrue(projectFile.exists());
+		assertTrue(projectFile.exists());
 
 		pressBack();
 
-		Assert.assertFalse(projectFile.exists());
+		assertFalse(projectFile.exists());
 
 		project = ProjectManager.getInstance().getCurrentProject();
 		onView(withText(NEW_PROJECT_NAME))
 				.check(matches(isDisplayed()));
-		Assert.assertEquals(NEW_PROJECT_NAME, project.getName());
+		assertEquals(NEW_PROJECT_NAME, project.getName());
 
 		onView(withText(R.string.default_project_background_name))
 				.perform(click());
@@ -171,13 +181,12 @@ public class ProjectOptionsTest {
 
 		intended(expectedPaintNewLookIntent);
 
-		onView(withText("Background (1)"))
-				.check(matches(isDisplayed()));
+		onRecyclerView().atPosition(0).onChildView(R.id.title_view)
+				.check(matches(withText(R.string.default_project_background_name)));
 	}
 
 	private Matcher<Intent> createLookFromPaintroid() throws IOException {
-		File tmpDir = new File(
-				Environment.getExternalStorageDirectory().getAbsolutePath(), "Pocket Code Test Temp");
+		File tmpDir = new File(Constants.CACHE_DIR.getAbsolutePath(), "Pocket Code Test Temp");
 		String lookFileName = "catroid_sunglasses.png";
 
 		Intents.init();
@@ -232,7 +241,8 @@ public class ProjectOptionsTest {
 
 		pressBack();
 
-		Assert.assertEquals(DESCRIPTION, project.getDescription());
+		project = ProjectManager.getInstance().getCurrentProject();
+		assertEquals(DESCRIPTION, project.getDescription());
 	}
 
 	@Test
@@ -244,7 +254,8 @@ public class ProjectOptionsTest {
 
 		pressBack();
 
-		Assert.assertEquals(NOTES_AND_CREDITS, project.getNotesAndCredits());
+		project = ProjectManager.getInstance().getCurrentProject();
+		assertEquals(NOTES_AND_CREDITS, project.getNotesAndCredits());
 	}
 
 	@Test
@@ -257,6 +268,7 @@ public class ProjectOptionsTest {
 		List<String> tagsList =
 				new ArrayList<>(Arrays.asList("Game", "Animation", "Tutorial"));
 
+		project = ProjectManager.getInstance().getCurrentProject();
 		project.setTags(tagsList);
 
 		openContextualActionModeOverflowMenu();
@@ -275,7 +287,7 @@ public class ProjectOptionsTest {
 
 	@Test
 	public void changeAspectRatio() {
-		Assert.assertEquals(ScreenModes.STRETCH, project.getScreenMode());
+		assertEquals(ScreenModes.STRETCH, project.getScreenMode());
 
 		onView(withId(R.id.project_options_aspect_ratio))
 				.perform(click());
@@ -285,7 +297,7 @@ public class ProjectOptionsTest {
 
 		pressBack();
 
-		Assert.assertEquals(ScreenModes.MAXIMIZE, project.getScreenMode());
+		assertEquals(ScreenModes.MAXIMIZE, project.getScreenMode());
 	}
 
 	@Test
@@ -300,17 +312,31 @@ public class ProjectOptionsTest {
 
 	@Test
 	public void saveExternal() throws IOException {
-		if (EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY.exists()) {
-			StorageOperations.deleteDir(EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY);
+		if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+
+			String doneToastText =
+					context.getString(
+							R.string.notification_save_project_to_external_storage_open,
+							EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY
+					) + "/" + PROJECT_NAME + CATROBAT_EXTENSION;
+
+			if (EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY.exists()) {
+				StorageOperations.deleteDir(EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY);
+			}
+
+			onView(withId(R.id.project_options_save_external))
+					.perform(ViewActions.scrollTo())
+					.perform(click());
+
+			onToast(withText(doneToastText))
+					.check(matches(isDisplayed()));
+			onView(isRoot()).perform(CustomActions
+					.wait(DURATION_WAIT_FOR_ZIP_FILE_IN_MILLISECONDS));
+
+			File externalProjectZip = new File(EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY,
+					project.getDirectory().getName() + CATROBAT_EXTENSION);
+			assertTrue(externalProjectZip.exists());
 		}
-
-		onView(withId(R.id.project_options_save_external))
-				.perform(ViewActions.scrollTo())
-				.perform(click());
-
-		File externalProjectZip = new File(EXTERNAL_STORAGE_ROOT_EXPORT_DIRECTORY,
-				project.getDirectory().getName() + CATROBAT_EXTENSION);
-		Assert.assertTrue(externalProjectZip.exists());
 	}
 
 	@Test
@@ -340,11 +366,11 @@ public class ProjectOptionsTest {
 				.check(matches(isDisplayed()));
 
 		File projectFile = new File(DEFAULT_ROOT_DIRECTORY, PROJECT_NAME);
-		Assert.assertTrue(projectFile.exists());
+		assertTrue(projectFile.exists());
 
 		onView(allOf(withId(android.R.id.button1), withText(R.string.yes)))
 				.perform(click());
 
-		Assert.assertFalse(projectFile.exists());
+		assertFalse(projectFile.exists());
 	}
 }
